Provide default settings, refactor, fix typos
