+++ annot8-components-base/src/main/java/io/annot8/components/base/components/AbstractResource.java
@@ -1,5 +1,13 @@
 package io.annot8.components.base.components;
 
-public abstract class AbstractResource {
+import io.annot8.core.components.Resource;
+
+
+/**
+ * Base class for resources.
+ *
+ * This is merely a convenience to allow just 'extends AbstractResource'.
+ */
+public abstract class AbstractResource extends AbstractComponent implements Resource {
 
 }
